+++ cli/codegen/generate_code.go
@@ -131,6 +131,14 @@ func main() {
 	if err != nil {
 		log.Errorf("error while generating file modes: %s", err)
 	}
+	err = generateFileModeFile(
+		"cli/create/builtin_templates/templates/vshard_cluster",
+		"cli/create/builtin_templates/static/vshard_cluster_template_filemodes_gen.go",
+		"VshardCluster",
+	)
+	if err != nil {
+		log.Errorf("error while generating file modes: %s", err)
+	}
 
 	if err = generateLuaCodeVar(); err != nil {
 		log.Errorf("error while generating lua code string variables: %s", err)